mm: compaction: get reference before non LRU movable folio isolation
Patch series "mm: convert to folio_isolate_movable()", v2.

Convert to the new folio_isolate_movable() to get rid of
isolate_movable_page().


This patch (of 5):

Non-LRU movable folio isolation will fail if it can't grab a reference in
isolate_movable_page(), so folio_get_nontail_page() could be called ahead
to unify the handling of non-LRU movable/LRU folio isolation a bit, this
is also prepare to convert isolate_movable_page() to take a folio.  Since
the reference count of the non-LRU movable folio is increased, a
folio_put() is needed whether the folio is isolated or not.

Link: https://lkml.kernel.org/r/20240829145456.2591719-1-wangkefeng.wang@huawei.com
Link: https://lkml.kernel.org/r/20240829145456.2591719-2-wangkefeng.wang@huawei.com
Signed-off-by: Kefeng Wang <wangkefeng.wang@huawei.com>
Reviewed-by: Vishal Moola (Oracle) <vishal.moola@gmail.com>
Cc: Baolin Wang <baolin.wang@linux.alibaba.com>
Cc: David Hildenbrand <david@redhat.com>
Cc: Matthew Wilcox <willy@infradead.org>
Cc: Zi Yan <ziy@nvidia.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>